backend/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
PyJWT==2.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
django-filter==2.4.0
python-dotenv
reportlab==3.6.11
drf-extra-fields==3.4.1
requests==2.22.0
